requirements.txt
flask
flask-mysqldb
PyJWT
Flask-SQLAlchemy